+++ resources/views/admin/comments/show.blade.php
@@ -25,7 +25,7 @@
                     <td>{{$comment->author}}</td>
                     <td>{{$comment->email}}</td>
                     <td>{{$comment->body}}</td>
-                    <td><a href="{{route('home.posts', $comment->post->id)}}">View Post</a></td>
+                    <td><a href="{{route('single.posts', $comment->post->slug)}}">View Post</a></td>
 
                     <td>
 
@@ -68,7 +68,7 @@
                     <td>
 
                         {{--<form method="post" action="/cms/public/posts">--}}
-                        {!! Form::open(['method'=>'DELETE', 'action'=>['PostCommentsController@destroy', $comment->id]]) !!}
+                        {!! Form::open(['method'=>'DELETE', 'action'=>['PostCommentsController@destroy', $comment->id],'onsubmit' => 'return ConfirmDelete()']) !!}
 
 
 
@@ -96,4 +96,18 @@
 
     @endif
 
+
+
 @stop
+<script>
+
+    function ConfirmDelete()
+    {
+        var x = confirm("Are you sure you want to delete?");
+        if (x)
+            return true;
+        else
+            return false;
+    }
+
+</script>
